remove ServiceException from DAO class

+++ lab3_folder/lab3_standalone/src/main/java/org/itmo/ws/jaxws/service/CarWebService.java
@@ -1,5 +1,6 @@
 package org.itmo.ws.jaxws.service;
 
+import org.itmo.ws.jaxws.exception.CarServiceFault;
 import org.itmo.ws.jaxws.exception.ServiceException;
 import org.itmo.ws.jaxws.model.PostgreSQLDAO;
 import org.itmo.ws.jaxws.model.Car;
@@ -25,19 +26,39 @@ public class CarWebService {
 
     @WebMethod(operationName = "create-new-car")
     public Long createNewCar(Car car) throws ServiceException{
+        if(!car.isValid()){
+            CarServiceFault fault = CarServiceFault.defaultInstance("Values are not provided for all fields");
+            throw new ServiceException("Values are not provided for all fields", fault);
+        }
+
         PostgreSQLDAO dao = new PostgreSQLDAO();
-        return dao.createNewCar(car);
+        Long id = dao.createNewCar(car);
+        if(id == -1L){
+            CarServiceFault fault = CarServiceFault.defaultInstance("Unable to add new record in table");
+            throw new ServiceException("Unable to add new record in table", fault);
+        }
+        return id;
     }
 
     @WebMethod(operationName = "delete-car")
     public Boolean deleteCar(Long id) throws ServiceException{
+        if(id == null || id <= 0){
+            CarServiceFault fault = CarServiceFault.defaultInstance("Provided incorrect id");
+            throw new ServiceException("Provided incorrect id", fault);
+        }
         PostgreSQLDAO dao = new PostgreSQLDAO();
         return dao.deleteCar(id);
     }
 
     @WebMethod(operationName = "update-car")
     public Boolean updateCar(Long id, Car car) throws ServiceException {
+        if(id == null || id <= 0){
+            CarServiceFault fault = CarServiceFault.defaultInstance("Provided incorrect id");
+            throw new ServiceException("Provided incorrect id", fault);
+        }
+
         PostgreSQLDAO dao = new PostgreSQLDAO();
-        return dao.updateCar(id,car);
+        Boolean result = dao.updateCar(id,car);
+        return result;
     }
 }
